api-gateway/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.7
uvicorn[standard]==0.34.3
pydantic==2.11.5
python-multipart==0.0.19

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database & Cache
redis==5.0.8
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# AI & ML Services
openai==1.59.6
pinecone-client==5.0.1
deepgram-sdk==3.10.3

# HTTP & WebSocket
httpx==0.28.1
websockets==14.1
aiofiles==24.1.0

# Monitoring & Logging
prometheus-client==0.21.1
structlog==24.5.0

# Utilities
orjson==3.10.15
email-validator==2.2.0

# Data processing
numpy>=2.0.0 